requirements.txt
pymongo
python-dotenv
numpy
pandas
python-dateutil
ipython
